pdf_pipeline/requirements.txt
# HTTP — para las REST APIs de Azure AI Foundry
requests>=2.32.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0

# Utilidades
tqdm>=4.66.0
rich>=13.7.0
tenacity>=8.3.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0

# Typing
typing-extensions>=4.11.0
